requirements.txt
flask
torch
torchvision
clip-by-openai
Pillow
requests